% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{affy.ivt.prep}
\alias{affy.ivt.prep}
\title{affymatrix ivt microarray preprocessing 'lite' version}
\usage{
affy.ivt.prep(array.fnames)
}
\arguments{
\item{array.fnames}{a character vector of array data file names}
}
\value{
a list of expressionMatrix and probe annotation info
}
\description{
affymatrix ivt microarray preprocessing 'lite' version
}
\author{
guangchun han
}
