% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{affy.Microarray.preprocessing}
\alias{affy.Microarray.preprocessing}
\title{internal fxn for affymatrix IVT microarray preprocessing}
\usage{
affy.Microarray.preprocessing(phenoData, Preprocessing = "mas5",
  PMA_cutoff = 0.1, auto.annot = T, symbol = F, ann.file = NULL,
  ann.col = 13, multiple2one = "Variance")
}
\arguments{
\item{phenoData}{a table derived from sdrf file which must have two columns called ArrayDataFile and State}

\item{Preprocessing}{prep method to use, possible values: mas5, dchip, rma, gcrma}

\item{PMA_cutoff}{percentage threshold of samples which failed the PMA calling test}

\item{auto.annot}{set to TURE for automatically annotation of the probes}

\item{symbol}{set to TURE to get gene symbol annotation instead of entrezID}

\item{ann.file}{if auto.annot = F, this file will be used to annote the probes}

\item{ann.col}{if auto.annot = F, this column of the ann.file will be used to annote the probes}

\item{multiple2one}{method for merge the probes corresponding to the same gene}
}
\value{
an expressionSet object
}
\description{
internal fxn for affymatrix IVT microarray preprocessing
}
\author{
dapeng, liang and guangchun, han
}
