% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{affy.st.prep}
\alias{affy.st.prep}
\title{affymatrix tilling microarray preprocessing}
\usage{
affy.st.prep(array.fnames)
}
\arguments{
\item{array.fnames}{a character vector of array data file names}
}
\value{
a list of expressionMatrix and probe annotation info
}
\description{
affymatrix tilling microarray preprocessing
}
\author{
guangchun han
}

