% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/microarray.R
\name{agilent44k.prep}
\alias{agilent44k.prep}
\title{agilent 4*44k microarray raw data preprocessing}
\usage{
agilent44k.prep(array.fnames, detect.rate = 0.2)
}
\arguments{
\item{array.fnames}{a character vector of array data file names}

\item{detect.rate}{PMA_cutoff analogue to affymatrix array}
}
\value{
a list of expressionMatrix and probe annotation info
}
\description{
agilent 4*44k microarray raw data preprocessing
}
\author{
guangchun han
}

