% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/microarray.R
\name{batchedDEG}
\alias{batchedDEG}
\title{calculate DEGs considering additional factors}
\usage{
batchedDEG(dat.preped, pheno.mtr, comp.cls.col, ...)
}
\arguments{
\item{dat.preped}{object returned by datPrep function}

\item{pheno.mtr}{this is the phenotype matrix which stores all phenotypes}

\item{comp.cls.col}{column number of the class in the pheno.mtr. DEGS will be calculated based on this class}

\item{...}{additional column numbers of factors that is used to group samples into sub-groups}
}
\value{
a list with class labels(useful for identify up and down regulated genes based on FC) and DEGs calcualted stored in the list
}
\description{
... for additional factors that is used to group samples into more precisable sub-groups
DEG will be calculated in each subgroups
}

