% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{batchedRP_filter}
\alias{batchedRP_filter}
\title{filter batchedRP DEG based on fold change and fdr}
\usage{
batchedRP_filter(deglist, fdr = 0.05, fc = 1.2)
}
\arguments{
\item{deglist}{obj returned by batchedRP}
}
\value{
a list with filtered DEGs and class label for each DEG calculation
}
\description{
this function is intended to be used accompany with batchedRP
}

