% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigcor.R
\name{bigcorXY}
\alias{bigcorXY}
\title{correlation on big matrix}
\usage{
bigcorXY(x, y, nblocks = NULL, verbose = TRUE, ...)
}
\arguments{
\item{x}{same as x in cor}

\item{y}{same as y in cor}

\item{nblocks}{number of blocks}
}
\description{
This fxn calculate expression on big matrix
}
\examples{

}
\author{
guangchun
}
\keyword{cor}
