% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineHTseq-count.r
\name{combineHTseq}
\alias{combineHTseq}
\title{combine htseq count result into a matrix}
\usage{
combineHTseq(x, sampleID = NA)
}
\arguments{
\item{x}{htseq count result root dir}

\item{sampleID}{optional sample names if provided, otherwise will use the subdirectory name}
}
\value{
a data frame
}
\description{
combine htseq count result into a matrix
}
\examples{

}
\author{
guangchun
}
\keyword{combineHTseq}
