% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{datPrep}
\alias{datPrep}
\title{internal fxn for array data preprocessing}
\usage{
datPrep(dat.mtr, featureName = NULL, sample.label, class1, class2)
}
\arguments{
\item{dat.mtr}{an expression data matrix with one column per sample; each gene as a row.}

\item{featureName}{optional gene name for the expression matrix(vector with length == nrow(dat.mtr)).}

\item{sample.label}{sample treament(condition) labels for each sample(vector with length == ncol(dat.mtr)).}

\item{class1}{class1 and class2 are strings indicating which groups are compared.}

\item{class2}{class1 and class2 are strings indicating which groups are compared.}
}
\value{
a list with 3 elements: dat for expression data matrix, cls for the coressponding sample labels, featureName is attached as also.
}
\description{
this function is used to extract and prepare the data matrix which will be used to do differential expression calculation.
if class1 is set to NULL, the returned value will be used for one class DEG analysis.
}
