% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/microarray.R
\name{DEG_filter}
\alias{DEG_filter}
\title{filter DEG based on fold change and fdr}
\usage{
DEG_filter(deglist, fdr = 0.05, fc = 1.2)
}
\arguments{
\item{deglist}{obj returned by batchedDEG}
}
\value{
a list with filtered DEGs and class label for each DEG calculation
}
\description{
this function is intended to be used accompany with batchedDEG
}

