% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/microarray.R
\name{diff.calc}
\alias{diff.calc}
\title{internal fxn for array data DEG calc}
\usage{
\method{diff}{calc}(dat.obj, diff.method, ...)
}
\arguments{
\item{dat.obj}{this is a list returned by dat.prep function.}

\item{diff.method}{possible values are: RP(for rankprod) limma sam}

\item{...}{further parameters passed to RP limma or sam algrithm}
}
\value{
a list with 2 tables: up&down regulated genes expression fold change table
}
\description{
internal fxn for array data DEG calc
}

