% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{diffOutput}
\alias{diffOutput}
\title{internal fxn for thresholding and output DE result}
\usage{
diffOutput(diff.calc.obj, FC = 1.5, fdr = 0.05, prefix = "")
}
\arguments{
\item{diff.calc.obj}{this is a list returned by diff.calc function.}

\item{FC}{fold change threshold}

\item{fdr}{false discovery rate threshold}

\item{prefix}{result file prefix string}
}
\value{
this fxn has no return value
}
\description{
internal fxn for thresholding and output DE result
}
