% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diskEaten.R
\name{diskEaten}
\alias{diskEaten}
\title{find the extraordinary large directories under the given dir}
\usage{
diskEaten(dir)
}
\arguments{
\item{dir}{the dir under which the contents will be checked}
}
\value{
this funciton return nothing, it prints out the result
}
\description{
to use this function under windows, download the du.exe binary from
https://technet.microsoft.com/en-us/sysinternals/du.aspx and  put it into PATH.
As this function use different underlying system API on win and linux,
the disk usage amount reported may not be the same.
}

