% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ebi.R
\name{ebi.affy.ivt.prep}
\alias{ebi.affy.ivt.prep}
\title{function to processing affymetrix ivt chip data(exon chip)}
\usage{
ebi.affy.ivt.prep(ae.flist)
}
\arguments{
\item{ae.flist}{is the object returned by ebi.fetch}
}
\description{
function to processing affymetrix ivt chip data(exon chip)
}

