% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebi.R
\name{ebi.affy.st.prep}
\alias{ebi.affy.st.prep}
\title{preprocessing affymetrix chip data}
\usage{
ebi.affy.st.prep(ae.flist)
}
\arguments{
\item{ae.flist}{object returned by ebi.fetch}
}
\description{
preprocessing affymetrix chip data
}

