% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebi.R
\name{ebi.agilent44k.prep}
\alias{ebi.agilent44k.prep}
\title{function to process agilent44k chip data
this function is part of the pipeline, the input is the value returned by ebi.fetch}
\usage{
ebi.agilent44k.prep(ae.flist)
}
\arguments{
\item{ae.flist}{object returned by ebi.fetch fxn}
}
\description{
function to process agilent44k chip data
this function is part of the pipeline, the input is the value returned by ebi.fetch
}

