% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ebi.R
\name{ebi.agilent80k.prep}
\alias{ebi.agilent80k.prep}
\title{function to process agilent80k chip data
this function actually calls ebi.agilent44k}
\usage{
ebi.agilent80k.prep(ae.flist)
}
\arguments{
\item{ae.flist}{object returned by ebi.fetch fxn}
}
\description{
function to process agilent80k chip data
this function actually calls ebi.agilent44k
}

