% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebi.R
\name{ebi.fetch}
\alias{ebi.fetch}
\title{fetch gene expression data from EBI arrayexpress}
\usage{
ebi.fetch(accession, path = getwd(), type = "full", extract = T,
  local = F, sourcedir = path, ...)
}
\arguments{
\item{accession}{Accession number of the data, like 'E-GEOD-6039' etc}

\item{path}{the directory where downloaded files will be stored in}

\item{type}{be one of raw/processed/full; for download the raw/processed/both data}

\item{extract}{default to T, otherwise the downloaded data will not be extracted}

\item{local}{local files to be read instead of downloading from the server}

\item{sourcedir}{if local is TRUE, files will be read from this dir}
}
\value{
return
}
\description{
this is a wrapper function which is used to download data from EBI arrayexpress
}

