% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ebi.R
\name{ebi.processed.prep}
\alias{ebi.processed.prep}
\title{function to read in processed data}
\usage{
ebi.processed.prep(ae.flist)
}
\arguments{
\item{accession}{the accession number of the dataset we want to analysis}
}
\description{
there are datasets that contains only processed data, in such condition, use this function to get a worked
object. There is a function called procset which does the same work in the ArrayExpress package. but it produces a
expressionSet object.
}

