% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_func.r
\name{enrich.test}
\alias{enrich.test}
\title{fisher's exact test and hypergeometric test for gene set}
\usage{
enrich.test(PathGenes, interestGenes, totalGenes, comet_exact_test = F,
  mutmatplot = T)
}
\arguments{
\item{PathGenes}{a vector of genes belonging to a pathway}

\item{interestGenes}{a vector of genes of interest}

\item{totalGenes}{a vector of background genes}

\item{comet_exact_test}{logical for comet exact test of mutual exclusive(requires package cometExactTest)}

\item{mutmatplot}{plot mutation mat for cometExactTest result}
}
\value{
a list
}
\description{
This fxn calculates statistics using fisher's exact test and hypergeometric test given a set of genes and gene set
}
\examples{

}
\author{
guangchun
}
\keyword{enrich.test}
