% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{fc.calc}
\alias{fc.calc}
\title{fc.calc for small sample DEG identification based on FC}
\usage{
fc.calc(dat.obj)
}
\arguments{
\item{dat.obj}{object returned by dat.prep function}
}
\value{
list with up&down regulated genes expression tables
}
\description{
fc.calc for small sample DEG identification based on FC
}

