% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findfactors.r
\name{findfactors}
\alias{findfactors}
\title{prime decomposition}
\usage{
findfactors(n)
}
\arguments{
\item{n}{number to decomposite}
}
\description{
This fxn decomposite a number into primes
}
\examples{

}
\author{
credit to https://rosettacode.org/wiki/Prime_decomposition#R
}
\keyword{decomposition}
\keyword{prime}
