% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/microarray.R
\name{gassigSearch}
\alias{gassigSearch}
\title{tackling with exon array gene annotation}
\usage{
gassigSearch(assignVec)
}
\arguments{
\item{assignVec}{the gene assignment vector}
}
\value{
NA or gene symbol
}
\description{
this function extrac gene symbo from the affy st arrays
}

