% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microarray.R
\name{genepix.prep}
\alias{genepix.prep}
\title{genepix array preprocessing(experimentail)}
\usage{
genepix.prep(array.fname)
}
\arguments{
\item{array.fnames}{a character vector of array data file names}
}
\value{
a list of expressionMatrix and probe annotation info
}
\description{
genepix array preprocessing(experimentail)
}
\author{
guangchun han
}

