% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/geo.R
\name{geo.agilent80k.prep}
\alias{geo.agilent80k.prep}
\title{agilent 4*44k microchip preprocessing method}
\usage{
geo.agilent80k.prep(geo.flist)
}
\arguments{
\item{geo.flist}{geo file list}
}
\description{
agilent 4*44k microchip preprocessing method
}

