% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/geo.R
\name{getMeta}
\alias{getMeta}
\title{phenotype data extractor of an geoquery object}
\usage{
getMeta(gse.obj, phe.num)
}
\arguments{
\item{gse.obj}{the geoquery object returned by GEOquery}

\item{phe.num}{specify the column number of the value}
}
\value{
phenotype information list with each element correponding to a sample
}
\description{
this fxn extract information from an object returned by GEOquery
}

