% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/geo.R
\name{getMtr}
\alias{getMtr}
\title{matrix extractor of an geoquery object}
\usage{
getMtr(gse.obj, col = 2)
}
\arguments{
\item{gse.obj}{the geoquery object returned by GEOquery}

\item{col}{specify the column number of expression value}
}
\value{
expression matrix with each column correponding to a sample, each row to a gene/probe
}
\description{
this fxn extract the expression matrix from an object returned by GEOquery
}

