% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{getSYMBLgpl}
\alias{getSYMBLgpl}
\title{get symbol from ncbi geo GPL table}
\usage{
getSYMBLgpl(probeIDs, gpl, sym.col)
}
\arguments{
\item{probeIDs}{probeId vector to be converted}

\item{gpl}{GPLxxx}

\item{sym.col}{as the column number containing symbols is not consistant, it's better to be specified by the user}
}
\value{
a vector of symbols with missing value as NA
}
\description{
get symbol from ncbi geo GPL table
}

