% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microarray.R
\name{grpFactor}
\alias{grpFactor}
\title{'combine' multiple factors}
\usage{
grpFactor(...)
}
\arguments{
\item{...}{factors to be passed in}

\item{sep}{the separater used for combine the factor levels, with '_' as the default}
}
\value{
a new factor with levels combined from each individual factor
}
\description{
this function will group factors of the same length into a 'combined' one
levels of inidividual factors will be combined together
}

