% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_gmt.R
\name{gsea.gmt.outputMatrix}
\alias{gsea.gmt.outputMatrix}
\title{output with each pathway as a column}
\usage{
gsea.gmt.outputMatrix(plist)
}
\arguments{
\item{x}{pathway list}
}
\value{
a matrix
}
\description{
output with each pathway as a column
}
\author{
ghan
}
