% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_gmt.R
\name{gsea.gmt.reOrganize}
\alias{gsea.gmt.reOrganize}
\title{re-organize the list returned by read.gsea.gmt}
\usage{
gsea.gmt.reOrganize(plist)
}
\arguments{
\item{x}{pathway list}
}
\value{
a list
}
\description{
This fxn accepts a list from read.gsea.gmt and reorganize it to a list with four elements: Names, Genes, nGene, urls
}
\examples{

}
\author{
guangchun
}
\keyword{gsea.gmt.reOrganize}
