% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanMouseSymbolMap.R
\name{human2mouse}
\alias{human2mouse}
\title{map gene symbols between mouse and human}
\usage{
human2mouse(hsymb)
}
\arguments{
\item{humansymbols}{}
}
\value{
a vector of corresponding mouse gene symbol
}
\description{
the function convert the symbol of human to the corresponding one of mouse
based on the record of homologen database in NCBI
}
\examples{
m1 <- c('APP', 'APOE')
human2mouse(m1)
}
\author{
guangchun
}
\keyword{human2mouse}
