% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microarray.R
\name{ilumi.preprocess}
\alias{ilumi.preprocess}
\title{internal fxn for illumina microarray preprocessing}
\usage{
ilumi.preprocess(dataFile, controlFile = NULL, arrayDesignFile = NULL,
  detection.p = 0.05, detectionRate = 0.5, norm.method = "ssn")
}
\arguments{
\item{dataFile}{this is the expression datafile exported from genomeStudio}

\item{controlFile}{control probe signal file. This file is optional, it is also exported form genomestudio for background correction}

\item{arrayDesignFile}{this is the `sdrf file`}

\item{detection.p}{detection pvalue threshold}

\item{detectionRate}{PMA_cutoff analogue to affymatrix array}

\item{norm.method}{normalization method}
}
\value{
a list of expressionMatrix, sample metaInfo, and probe annotation info
}
\description{
internal fxn for illumina microarray preprocessing
}
\author{
guangchun han
}

