% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2matrix.R
\name{list2matrix}
\alias{list2matrix}
\title{A list to matrix transform function}
\usage{
list2matrix(x)
}
\arguments{
\item{x}{a list of vectors}
}
\value{
a data frame
}
\description{
This fxn accepts a list and transform it to a matrix if possible
missing entries will be filled with NA in the returned dataframe
}
\examples{
tmp <- list(x=c(2,8,6,4,3),
            b=c(6,33,5,64,85,90),
            d=c('ks','foo','bar'))
list2matrix(tmp)
}
\author{
guangchun
}
\keyword{list2matrix}

