% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccTable.R
\name{list2OccTable1}
\alias{list2OccTable1}
\title{convert a named list to an Ocurrence table(emerged 1, absent 0), this one is speedy}
\usage{
list2OccTable1(lst)
}
\arguments{
\item{lst}{a deg list}
}
\value{
a matrix with 0s and 1s
}
\description{
this function convert a names list  with length n to a matrix with dim x*n
x is the total number of unique elements stored in the list
}
