% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microarray.R
\name{list2table}
\alias{list2table}
\title{rearrange degs.list to degs table}
\usage{
list2table(degs.list)
}
\description{
assume that degs.list is a list of degs. each element is a list of length 2:up regulated genes&down regulated genes
}

