% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/list2matrix.R
\name{maxlen}
\alias{maxlen}
\title{internal fxn used in list2matrix}
\usage{
maxlen(x)
}
\arguments{
\item{x}{a list of vectors}
}
\description{
return the length of the vector element with max length
}
\keyword{length}

