% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanMouseSymbolMap.R
\name{mouse2human}
\alias{mouse2human}
\title{map gene symbols between mouse and human}
\usage{
mouse2human(msymb)
}
\arguments{
\item{mousesymbols}{}
}
\value{
a vector of corresponding human gene symbol
}
\description{
the function convert the symbol of mouse to the corresponding one of human
based on the record of homologen database in NCBI
}
\examples{
m1 <- c('App', 'Apoe')
mouse2human(m1)
}
\author{
guangchun
}
\keyword{mouse2human}

