% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagram.R
\name{myVennDiagram}
\alias{myVennDiagram}
\title{VennDiagam with color and font set}
\usage{
myVennDiagram(...)
}
\description{
It's based on VennDiagram package with color, fontfamily set.
Color scheme was chosen from color brewer. 
It takes the same parameter as venn.diagram
You can override the paramters by specifiying them again
}

