% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{nimblegene.prep}
\alias{nimblegene.prep}
\title{nimblegene array preprocessing(experimentail)}
\usage{
nimblegene.prep(array.fnames, detect.rate = 0.2)
}
\arguments{
\item{array.fnames}{a character vector of array data file names}
}
\value{
a list of expressionMatrix and probe annotation info as well as a sdrf file
}
\description{
nimblegene array preprocessing(experimentail)
}
\author{
guangchun han
}
