% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_gmt.R
\name{read.gsea.gmt}
\alias{read.gsea.gmt}
\title{read in gmt gene set file from gsea}
\usage{
read.gsea.gmt(f)
}
\arguments{
\item{x}{gmt file name}
}
\value{
a list
}
\description{
This fxn accepts a gmt file and transform it to a list
}
\examples{

}
\author{
guangchun
}
\keyword{read.gsea.gmt}
