% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameOrder.R
\name{sameOrder}
\alias{sameOrder}
\title{Check if two character vectors has the same element order}
\usage{
sameOrder(x, y)
}
\arguments{
\item{x}{vector, must be a character vector with no duplicated elements}

\item{y}{another vector which is used to be compared against, must be a character vector with no duplicated elements too}
}
\value{
a list with two components: status and position
}
\description{
Usually, we have to check taht the col names(sample name) of
an expression matrix is in the same order with the sample names
listed in the metainfor data matrix, this function is intended
for this purpose
}
\examples{

x<-sample(letters,6)
y<-sample(x, 6)

samOrder(x,x)
samOrder(x,y)

}
\author{
guangchun hanguangchunATgmail.com
}
\keyword{samOrder}
