% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNCBI.R
\name{searchNCBI}
\alias{searchNCBI}
\title{search NCBI pubmed for wanderful papers}
\usage{
searchNCBI(keyword, db = "pubmed", out = NULL)
}
\arguments{
\item{keyword}{}
}
\value{
a data frame
}
\description{
search NCBI pubmed for wanderful papers
}
\author{
guangchun
}
\keyword{searchNCBI}
