% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/geo.R
\name{soft2exp.pval}
\alias{soft2exp.pval}
\title{expression matrix generator (with detection pvalue)}
\usage{
soft2exp.pval(AccNum, detectp.col = 3, exp.col = 2, detectp = 0.05,
  detectionRate = 0.5, log2 = F)
}
\arguments{
\item{AccNum}{accession number}

\item{detectp.col}{column number of detection p value}

\item{exp.col}{column number of expression value}

\item{detectionRate}{threshold of sample detection rate for each gene}

\item{log2}{set to T for logrithm transformation}
}
\value{
a list with expmatrix probeID symbols and entrezID
}
\description{
given an accession number(gseXXXX), this fxn download and extract the expression matrix
}

