% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/geo.R
\name{soft2exp}
\alias{soft2exp}
\title{expression matrix generator}
\usage{
soft2exp(AccNum, filename = NULL, exp.col = 2, log2 = F)
}
\arguments{
\item{AccNum}{accession number}

\item{filename}{read data from the local data instead of the remote server}

\item{exp.col}{column number of expression value}

\item{log2}{set to T for logrithm transformation}
}
\value{
a list with expmatrix probeID symbols and entrezID
}
\description{
given an accession number(gseXXXX), this fxn download and extract the expression matrix
}

