% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/whichmaxlen.R
\name{whichmaxlen}
\alias{whichmaxlen}
\title{inner fxn used in list2matrix}
\usage{
whichmaxlen(x)
}
\arguments{
\item{x}{a list of vectors}
}
\description{
return the index of the element with max vector length in a list of vectors.
if there are many longest vectors,return a vector of their indice
}
\keyword{index}

