% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNCBI.R
\name{searchNCBI}
\alias{searchNCBI}
\title{search NCBI pubmed for wanderful papers}
\usage{
searchNCBI(keyword, db = "pubmed", out = NULL)
}
\arguments{
\item{keyword}{the query statement used to search in ncbi pubmed website}
}
\value{
a data frame with IF2017 included
}
\description{
search NCBI pubmed for wanderful papers
}
\author{
hanguangchun@gmail.com
}
\keyword{searchNCBI}
